[Refactor] Split virtual_uart host application in multiple files
